refactor: split review action boundaries
