+++ src/hooks/useAppReviewActions.ts
@@ -1,343 +1,16 @@
-﻿import type { Dispatch, MutableRefObject, SetStateAction } from 'react';
-import {
-  createComment,
-  createReview,
-  deleteComment,
-  deleteReview,
-  toggleReviewLike,
-  updateComment,
-  updateReview,
-  uploadReviewImage,
-} from '../api/reviewsClient';
-import { countCommentsInThread, toReviewSummary } from '../lib/reviews';
-import { useAppPageRuntimeStore } from '../store/app-page-runtime-store';
-import { useReviewUIStore } from '../store/review-ui-store';
-import type { RouteStateCommitOptions } from './useAppRouteState';
-import type {
-  Comment,
-  DrawerState,
-  MyPageResponse,
-  Place,
-  Review,
-  ReviewMood,
-  SessionUser,
-  Tab,
-} from '../types';
+import { useAppReviewCommentActions } from './useAppReviewCommentActions';
+import { useAppReviewCrudActions } from './useAppReviewCrudActions';
+import { useAppReviewLikeActions } from './useAppReviewLikeActions';
+import type { UseAppReviewActionsParams } from './useAppReviewActions.types';
 
-type SetState<T> = Dispatch<SetStateAction<T>>;
-type HistoryMode = 'push' | 'replace';
-
-interface UseAppReviewActionsParams {
-  activeTab: Tab;
-  sessionUser: SessionUser | null;
-  selectedPlace: Place | null;
-  reviews: Review[];
-  selectedPlaceReviews: Review[];
-  myPage: MyPageResponse | null;
-  activeCommentReviewId: string | null;
-  highlightedReviewId: string | null;
-  setSelectedPlaceReviews: SetState<Review[]>;
-  setReviews: SetState<Review[]>;
-  setMyPage: SetState<MyPageResponse | null>;
-  setNotice: (notice: string | null) => void;
-  goToTab: (nextTab: Tab, historyMode?: HistoryMode) => void;
-  commitRouteState: (
-    nextState: { tab: Tab; placeId: string | null; festivalId: string | null; drawerState: DrawerState },
-    historyMode?: HistoryMode,
-    options?: RouteStateCommitOptions,
-  ) => void;
-  refreshMyPageForUser: (user: SessionUser | null, force?: boolean) => Promise<MyPageResponse | null>;
-  patchReviewCollections: (reviewId: string, updater: (review: Review) => Review) => void;
-  upsertReviewCollections: (review: Review) => void;
-  placeReviewsCacheRef: MutableRefObject<Record<string, Review[]>>;
-  handleCloseReviewComments: () => void;
-  syncReviewComments: (reviewId: string, comments: Comment[]) => void;
-  clearReviewComments: (reviewId: string) => void;
-  formatErrorMessage: (error: unknown) => string;
-}
-
-export function useAppReviewActions({
-  activeTab,
-  sessionUser,
-  selectedPlace,
-  reviews,
-  selectedPlaceReviews,
-  myPage,
-  activeCommentReviewId,
-  highlightedReviewId,
-  setSelectedPlaceReviews,
-  setReviews,
-  setMyPage,
-  setNotice,
-  goToTab,
-  commitRouteState,
-  refreshMyPageForUser,
-  patchReviewCollections,
-  upsertReviewCollections,
-  placeReviewsCacheRef,
-  handleCloseReviewComments,
-  syncReviewComments,
-  clearReviewComments,
-  formatErrorMessage,
-}: UseAppReviewActionsParams) {
-  const setReviewSubmitting = useAppPageRuntimeStore((state) => state.setReviewSubmitting);
-  const setReviewError = useAppPageRuntimeStore((state) => state.setReviewError);
-  const setCommentSubmittingReviewId = useAppPageRuntimeStore((state) => state.setCommentSubmittingReviewId);
-  const setCommentMutatingId = useAppPageRuntimeStore((state) => state.setCommentMutatingId);
-  const setDeletingReviewId = useAppPageRuntimeStore((state) => state.setDeletingReviewId);
-  const setReviewLikeUpdatingId = useAppPageRuntimeStore((state) => state.setReviewLikeUpdatingId);
-  const setHighlightedReviewId = useReviewUIStore((state) => state.setHighlightedReviewId);
-
-  async function handleCreateReview(payload: { stampId: string; body: string; mood: ReviewMood; file: File | null }) {
-    if (!sessionUser || !selectedPlace) {
-      goToTab('my');
-      return;
-    }
-
-    setReviewSubmitting(true);
-    setReviewError(null);
-    try {
-      let imageUrl: string | null = null;
-      if (payload.file) {
-        const uploaded = await uploadReviewImage(payload.file);
-        imageUrl = uploaded.url;
-      }
-      const createdReview = await createReview({
-        placeId: selectedPlace.id,
-        stampId: payload.stampId,
-        body: payload.body.trim(),
-        mood: payload.mood,
-        imageUrl,
-      });
-      upsertReviewCollections(createdReview);
-      await refreshMyPageForUser(sessionUser);
-      setNotice('피드를 남겼어요. 이제 다른 장소도 이어서 둘러볼 수 있어요.');
-      commitRouteState(
-        {
-          tab: 'map',
-          placeId: selectedPlace.id,
-          festivalId: null,
-          drawerState: 'full',
-        },
-        'replace',
-      );
-    } catch (error) {
-      setReviewError(formatErrorMessage(error));
-    } finally {
-      setReviewSubmitting(false);
-    }
-  }
-
-  async function handleUpdateReview(reviewId: string, payload: { body: string; mood: ReviewMood; file?: File | null; removeImage?: boolean }) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('피드를 수정하려면 먼저 로그인해 주세요.');
-      return;
-    }
-
-    let imageUrl: string | null | undefined;
-    if (payload.file) {
-      const uploaded = await uploadReviewImage(payload.file);
-      imageUrl = uploaded.url;
-    } else if (payload.removeImage) {
-      imageUrl = null;
-    }
-
-    const updatedReview = await updateReview(reviewId, {
-      body: payload.body.trim(),
-      mood: payload.mood,
-      imageUrl,
-    });
-    const summarizedReview = toReviewSummary(updatedReview);
-    patchReviewCollections(reviewId, () => summarizedReview);
-    setMyPage((current) => {
-      if (!current) {
-        return current;
-      }
-      return {
-        ...current,
-        reviews: current.reviews.map((review) => (review.id === reviewId ? summarizedReview : review)),
-        comments: current.comments.map((comment) => (
-          comment.reviewId === reviewId
-            ? { ...comment, reviewBody: updatedReview.body }
-            : comment
-        )),
-      };
-    });
-    setNotice('피드를 수정했어요.');
-  }
-
-  async function handleCreateComment(reviewId: string, body: string, parentId?: string) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('댓글을 남기려면 먼저 로그인해 주세요.');
-      return;
-    }
-
-    setCommentSubmittingReviewId(reviewId);
-    try {
-      const updatedComments = await createComment(reviewId, { body, parentId: parentId ?? null });
-      syncReviewComments(reviewId, updatedComments);
-      patchReviewCollections(reviewId, (review) => ({
-        ...review,
-        commentCount: countCommentsInThread(updatedComments),
-      }));
-    } catch (error) {
-      setNotice(formatErrorMessage(error));
-    } finally {
-      setCommentSubmittingReviewId(null);
-    }
-  }
-
-  async function handleUpdateComment(reviewId: string, commentId: string, body: string) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('댓글을 수정하려면 먼저 로그인해 주세요.');
-      return;
-    }
-
-    setCommentMutatingId(commentId);
-    try {
-      const updatedComments = await updateComment(reviewId, commentId, { body });
-      syncReviewComments(reviewId, updatedComments);
-      patchReviewCollections(reviewId, (review) => ({
-        ...review,
-        commentCount: countCommentsInThread(updatedComments),
-      }));
-      if (activeTab === 'my') {
-        await refreshMyPageForUser(sessionUser, true);
-      }
-    } catch (error) {
-      setNotice(formatErrorMessage(error));
-    } finally {
-      setCommentMutatingId(null);
-    }
-  }
-
-  async function handleDeleteComment(reviewId: string, commentId: string) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('댓글을 삭제하려면 먼저 로그인해 주세요.');
-      return;
-    }
-
-    setCommentMutatingId(commentId);
-    try {
-      const updatedComments = await deleteComment(reviewId, commentId);
-      syncReviewComments(reviewId, updatedComments);
-      patchReviewCollections(reviewId, (review) => ({
-        ...review,
-        commentCount: countCommentsInThread(updatedComments),
-      }));
-      if (activeTab === 'my') {
-        await refreshMyPageForUser(sessionUser, true);
-      }
-    } catch (error) {
-      setNotice(formatErrorMessage(error));
-    } finally {
-      setCommentMutatingId(null);
-    }
-  }
-
-  async function handleDeleteReview(reviewId: string) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('피드를 삭제하려면 먼저 로그인해 주세요.');
-      return;
-    }
-    if (!window.confirm('이 피드를 삭제할까요?')) {
-      return;
-    }
-
-    setDeletingReviewId(reviewId);
-    try {
-      await deleteReview(reviewId);
-      clearReviewComments(reviewId);
-      setReviews((current) => current.filter((review) => review.id !== reviewId));
-      setSelectedPlaceReviews((current) => current.filter((review) => review.id !== reviewId));
-      for (const placeId of Object.keys(placeReviewsCacheRef.current)) {
-        placeReviewsCacheRef.current[placeId] = placeReviewsCacheRef.current[placeId].filter((review) => review.id !== reviewId);
-      }
-      setMyPage((current) => {
-        if (!current) {
-          return current;
-        }
-        return {
-          ...current,
-          reviews: current.reviews.filter((review) => review.id !== reviewId),
-          comments: current.comments.filter((comment) => comment.reviewId !== reviewId),
-          stats: {
-            ...current.stats,
-            reviewCount: Math.max(0, current.stats.reviewCount - 1),
-          },
-        };
-      });
-      if (activeCommentReviewId === reviewId) {
-        handleCloseReviewComments();
-      }
-      if (highlightedReviewId === reviewId) {
-        setHighlightedReviewId(null);
-      }
-      setNotice('피드를 삭제했어요.');
-      if (activeTab === 'my') {
-        await refreshMyPageForUser(sessionUser, true);
-      }
-    } catch (error) {
-      setNotice(formatErrorMessage(error));
-    } finally {
-      setDeletingReviewId(null);
-    }
-  }
-
-  async function handleToggleReviewLike(reviewId: string) {
-    if (!sessionUser) {
-      goToTab('my');
-      setNotice('좋아요를 누르려면 먼저 로그인해 주세요.');
-      return;
-    }
-
-    const targetReview = reviews.find((review) => review.id === reviewId)
-      ?? selectedPlaceReviews.find((review) => review.id === reviewId)
-      ?? myPage?.reviews.find((review) => review.id === reviewId)
-      ?? null;
-    const previousLikedByMe = targetReview?.likedByMe ?? false;
-    const previousLikeCount = targetReview?.likeCount ?? 0;
-    const optimisticLikedByMe = !previousLikedByMe;
-    const optimisticLikeCount = Math.max(0, previousLikeCount + (optimisticLikedByMe ? 1 : -1));
-
-    patchReviewCollections(reviewId, (review) => ({
-      ...review,
-      likeCount: optimisticLikeCount,
-      likedByMe: optimisticLikedByMe,
-    }));
-
-    setReviewLikeUpdatingId(reviewId);
-    try {
-      const result = await toggleReviewLike(reviewId);
-      patchReviewCollections(reviewId, (review) => ({
-        ...review,
-        likeCount: result.likeCount,
-        likedByMe: result.likedByMe,
-      }));
-    } catch (error) {
-      patchReviewCollections(reviewId, (review) => ({
-        ...review,
-        likeCount: previousLikeCount,
-        likedByMe: previousLikedByMe,
-      }));
-      setNotice(formatErrorMessage(error));
-    } finally {
-      setReviewLikeUpdatingId(null);
-    }
-  }
+export function useAppReviewActions(params: UseAppReviewActionsParams) {
+  const reviewCrudActions = useAppReviewCrudActions(params);
+  const commentActions = useAppReviewCommentActions(params);
+  const reviewLikeActions = useAppReviewLikeActions(params);
 
   return {
-    handleCreateReview,
-    handleUpdateReview,
-    handleCreateComment,
-    handleUpdateComment,
-    handleDeleteComment,
-    handleDeleteReview,
-    handleToggleReviewLike,
+    ...reviewCrudActions,
+    ...commentActions,
+    ...reviewLikeActions,
   };
 }
